Move vec_box to its own module
